% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindexdecomp-package.R
\docType{package}
\name{cindexdecomp-package}
\alias{cindexdecomp}
\alias{cindexdecomp-package}
\title{cindexdecomp: C-Index Decomposition for Survival Models}
\description{
Decomposes concordance indices for survival models into Event-Event and Event-Censored components to expose performance masking in highly censored data. Supports Harrell's C, Uno's inverse-probability-of-censoring-weighted C, time-truncated C, and user-supplied pair weightings through a single engine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lainsm/cindexdecomp}
  \item \url{https://Lainsm.github.io/cindexdecomp}
  \item Report bugs at \url{https://github.com/Lainsm/cindexdecomp/issues}
}

}
\author{
\strong{Maintainer}: Malik J. Lainsbury \email{mjl221@cam.ac.uk}

}
\keyword{internal}
