% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plot the Censoring Illusion Simulation}
\usage{
plot_simulation(
  sim_df,
  col_ee = "#DC267F",
  col_ec = "#648FFF",
  background = "#282a36"
)
}
\arguments{
\item{sim_df}{A data frame output from \code{simulate_censoring()}}

\item{col_ee}{Character. Colour for C_ee line.
Default is IBM magenta "#DC267F".}

\item{col_ec}{Character. Colour for C_ec line.
Default is IBM blue "#648FFF".}

\item{background}{Character. Plot background colour.
Default is Dracula "#282a36".}
}
\value{
A ggplot2 object
}
\description{
Visualises the result of \code{simulate_censoring()} as a line chart
showing how C_ee, C_ec and Global C-Index change as artificial
censoring increases. The shaded region between Global C-Index and
C_ee represents the "Censoring Illusion Zone".
}
\examples{
set.seed(42)
time   <- rexp(200, rate = 0.1)
status <- rbinom(200, 1, 0.6)
risk   <- rnorm(200)

sim <- simulate_censoring(time, status, risk)
plot_simulation(sim)

}
